app/src/test/java/com/cshawn/codes/leetcodes/everyday/Q22Test.java
package com.cshawn.codes.leetcodes.everyday;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;
import java.util.stream.Stream;
import static org.junit.jupiter.api.Assertions.assertArrayEquals;

/**
 * @author C.Shawn
 * @date 2021/10/27 5:33 下午
 */
class Q22Test {
    static class DataArgumentsProvider implements ArgumentsProvider {
        @Override
        public Stream<? extends Arguments> provideArguments(ExtensionContext context) throws Exception {
            return Stream.of(
                    Arguments.of(1, new String[]{"()"}),
                    Arguments.of(2, new String[]{"(())", "()()"}),
                    Arguments.of(3, new String[]{"((()))", "(()())", "(())()", "()(())", "()()()"})
            );
        }
    }

    @ParameterizedTest
    @ArgumentsSource(DataArgumentsProvider.class)
    void generateParenthesis(int n, String[] result) {
        assertArrayEquals(result, new Q22().generateParenthesis(n).toArray());
    }
}